LayoutTests/http/tests/referrer-policy-anchor/origin-when-cross-origin/cross-origin-http-http-expected.txt

Tests the behavior of origin-when-cross-origin referrer policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is expected
PASS successfullyParsed is true

TEST COMPLETE
